+++ __mocks__/mapDefinitions/FullTranscriptMapDefinition.ts
@@ -30,7 +30,7 @@ ns0:Root:
       $for(/ns0:Root/CumulativeExpression/Population/State):
         State:
           Name: Name
-          SexRatio: divide(count(County/Person/Sex/Male), count(County/Person/Sex/Female))
+          SexRatio: string(divide(count(County/Person/Sex/Male), count(County/Person/Sex/Female)))
   ConditionalMapping:
     ItemPrice: /ns0:Root/ConditionalMapping/ItemPrice
     ItemQuantity: /ns0:Root/ConditionalMapping/ItemQuantity
@@ -49,10 +49,10 @@ ns0:Root:
           Distance: Distance
           Duration: Duration
   ConditionalLooping:
-    CategorizedCatalog:
-      # Following loop with 3 $if can be coded alternatively by including complex expressions inside selector in $for statements for-example: Product[is-equal(substring(SKU, 1, 2), "11")].
-      # This version is simplified to what Data Mapper front end will generate. Map code can be manually authored with complex selector expressions for optimization or ease of read.
-      $for(/ns0:Root/ConditionalLooping/FlatterCatalog/ns0:Product):
+    # Following loop with 3 $if can be coded alternatively by including complex expressions inside selector in $for statements for-example: Product[is-equal(substring(SKU, 1, 2), "11")].
+    # This version is simplified to what Data Mapper front end will generate. Map code can be manually authored with complex selector expressions for optimization or ease of read.
+    $for(/ns0:Root/ConditionalLooping/FlatterCatalog/ns0:Product):
+      CategorizedCatalog:
         $if(is-equal(substring(SKU, 1, 2), "11")):
           PetProduct:
             Name: Name
@@ -88,20 +88,20 @@ ns0:Root:
             WindSpeed: ./@WindSpeed
             Temperature: ./@Temperature
   NameValueTransforms:
-    # FlatterCatalog: # TODO: Conditionals (at least property ones) need some help when within loops
-      # $for(/ns0:Root/NameValueTransforms/Catalog/Product):
-        # $for(Field):
-          # Price:
-            # $if(is-equal(Name, "Pen")):
-              # Pen: Value
-            # $if(is-equal(Name, "Pencil")):
-              # Pencil: Value
-            # $if(is-equal(Name, "Notebook")):
-              # NoteBook: Value
-            # $if(is-equal(Name, "Bag")):
-              # Bag: Value
-            # $if(is-equal(Name, "Others")):
-              # Others: Value
+    FlatterCatalog:
+      $for(/ns0:Root/NameValueTransforms/Catalog/Product):
+        $for(Field):
+          Price:
+            $if(is-equal(Name, "Pen")):
+              Pen: Value
+            $if(is-equal(Name, "Pencil")):
+              Pencil: Value
+            $if(is-equal(Name, "Notebook")):
+              NoteBook: Value
+            $if(is-equal(Name, "Bag")):
+              Bag: Value
+            $if(is-equal(Name, "Others")):
+              Others: Value
     PO_Status:
       ShipDate: /ns0:Root/NameValueTransforms/PurchaseOrderStatus/DateShipped
       ShippedVia: '"Air"'