+++ libs/designer/src/lib/core/actions/bjsworkflow/initialize.ts
@@ -120,10 +120,10 @@ export const getInputParametersFromManifest = (
 
     primaryInputParametersInArray = updateParameterWithValues(
       'inputs.$',
-      getInputsValueFromDefinitionForManifest(inputsLocation ?? ['inputs'], operationData),
+      getInputsValueFromDefinitionForManifest(inputsLocation ?? ['inputs'], manifest, operationData, primaryInputParametersInArray),
       '',
       primaryInputParametersInArray,
-      !inputsLocation || !!inputsLocation.length /* createInvisibleParameter */,
+      (!inputsLocation || !!inputsLocation.length) && !manifest.properties.inputsLocationSwapMap /* createInvisibleParameter */,
       false /* useDefault */
     );
   } else {