+++ libs/designer/src/lib/core/actions/bjsworkflow/serializer.ts
@@ -22,8 +22,11 @@ import { UIConstants } from '@microsoft/designer-ui';
 import { getIntl } from '@microsoft/intl-logic-apps';
 import type { Segment } from '@microsoft/parsers-logic-apps';
 import { create, removeConnectionPrefix, cleanIndexedValue, isAncestorKey, parseEx, SegmentType } from '@microsoft/parsers-logic-apps';
-import type { OperationManifest, SubGraphDetail } from '@microsoft/utils-logic-apps';
+import type { LocationSwapMap, OperationManifest, SubGraphDetail } from '@microsoft/utils-logic-apps';
 import {
+  clone,
+  deleteObjectProperty,
+  getObjectPropertyValue,
   equals,
   isNullOrUndefined,
   safeSetObjectPropertyValue,
@@ -205,7 +208,7 @@ const serializeManifestBasedOperation = async (rootState: RootState, operationId
   const nodeSettings = rootState.operations.settings[operationId] ?? {};
   const inputPathValue = serializeOperationParameters(inputsToSerialize, manifest);
   const hostInfo = serializeHost(operationId, manifest, rootState);
-  const inputs = hostInfo !== undefined ? { ...inputPathValue, ...hostInfo } : inputPathValue;
+  const inputs = hostInfo !== undefined ? mergeHostWithInputs(hostInfo, inputPathValue) : inputPathValue;
   const operationFromWorkflow = rootState.workflow.operations[operationId];
   const runAfter = isRootNode(operationId, rootState.workflow.nodesMetadata)
     ? undefined
@@ -281,12 +284,10 @@ export interface SerializedParameter extends ParameterInfo {
 
 const getOperationInputsToSerialize = (rootState: RootState, operationId: string): SerializedParameter[] => {
   const idReplacements = rootState.workflow.idReplacements;
-  return getOperationInputParameters(rootState, operationId)
-    .filter((input) => !input.info.serialization?.skip)
-    .map((input) => ({
-      ...input,
-      value: parameterValueToString(input, true /* isDefinitionValue */, idReplacements),
-    }));
+  return getOperationInputParameters(rootState, operationId).map((input) => ({
+    ...input,
+    value: parameterValueToString(input, true /* isDefinitionValue */, idReplacements),
+  }));
 };
 
 const serializeOperationParameters = (inputs: SerializedParameter[], manifest: OperationManifest): Record<string, any> => {
@@ -299,7 +300,7 @@ const serializeOperationParameters = (inputs: SerializedParameter[], manifest: O
     parametersValue = property === '[*]' ? [parametersValue] : { [property]: parametersValue };
   }
 
-  return parametersValue;
+  return swapInputsLocationIfNeeded(parametersValue, manifest.properties.inputsLocationSwapMap);
 };
 
 export const constructInputValues = (key: string, inputs: SerializedParameter[], encodePathComponents: boolean): any => {
@@ -314,14 +315,34 @@ export const constructInputValues = (key: string, inputs: SerializedParameter[],
     }
     return result !== undefined ? result : rootParameter.required ? null : undefined;
   } else {
-    const descendantParameters = inputs.filter((item) => isAncestorKey(item.parameterKey, key));
-    for (const serializedParameter of descendantParameters) {
-      let parameterValue = getJSONValueFromString(serializedParameter.value, serializedParameter.type);
-      if (encodePathComponents) {
-        const encodeCount = getEncodeValue(serializedParameter.info.encode ?? '');
-        parameterValue = encodePathValue(parameterValue, encodeCount);
+    const propertyNameParameters: SerializedParameter[] = [];
+    const serializedParameters = inputs
+      .filter((item) => isAncestorKey(item.parameterKey, key))
+      .map((descendantParameter) => {
+        let parameterValue = getJSONValueFromString(descendantParameter.value, descendantParameter.type);
+        if (encodePathComponents) {
+          const encodeCount = getEncodeValue(descendantParameter.info.encode ?? '');
+          parameterValue = encodePathValue(parameterValue, encodeCount);
+        }
+
+        const serializedParameter = { ...descendantParameter, value: parameterValue };
+        if (descendantParameter.info.serialization?.property) {
+          propertyNameParameters.push(serializedParameter);
+        }
+
+        return serializedParameter;
+      });
+
+    for (const serializedParameter of serializedParameters) {
+      if (!propertyNameParameters.find((param) => param.parameterKey === serializedParameter.parameterKey)) {
+        let parameterKey = serializedParameter.parameterKey;
+        for (const propertyNameParameter of propertyNameParameters) {
+          const propertyName = propertyNameParameter.info.serialization?.property?.name as string;
+          parameterKey = parameterKey.replace(propertyName, propertyNameParameter.value);
+        }
+
+        result = serializeParameterWithPath(result, serializedParameter.value, key, { ...serializedParameter, parameterKey });
       }
-      result = serializeParameterWithPath(result, parameterValue, key, serializedParameter);
     }
   }
 
@@ -404,9 +425,30 @@ const serializeParametersFromSwagger = async (
   return parameterInputs;
 };
 
+const swapInputsLocationIfNeeded = (parametersValue: any, swapMap: LocationSwapMap[] | undefined): any => {
+  if (!swapMap?.length) {
+    return parametersValue;
+  }
+
+  let finalValue = clone(parametersValue);
+  for (const { source, target } of swapMap) {
+    const value = getObjectPropertyValue(parametersValue, source);
+    deleteObjectProperty(finalValue, source);
+    finalValue = !target.length ? { ...finalValue, ...value } : safeSetObjectPropertyValue(finalValue, target, value);
+  }
+
+  return finalValue;
+};
+
 //#endregion
 
 //#region Host Serialization
+interface ApiManagementConnectionInfo {
+  apiManagement: {
+    connection: string;
+  };
+}
+
 interface FunctionConnectionInfo {
   function: {
     connectionName: string;
@@ -425,7 +467,7 @@ const serializeHost = (
   nodeId: string,
   manifest: OperationManifest,
   rootState: RootState
-): FunctionConnectionInfo | ServiceProviderConnectionConfigInfo | undefined => {
+): FunctionConnectionInfo | ApiManagementConnectionInfo | ServiceProviderConnectionConfigInfo | undefined => {
   if (!manifest.properties.connectionReference) {
     return undefined;
   }
@@ -440,6 +482,12 @@ const serializeHost = (
       return {
         function: {
           connectionName: referenceKey,
+        },
+      };
+    case ConnectionReferenceKeyFormat.ApiManagement:
+      return {
+        apiManagement: {
+          connection: referenceKey,
         },
       };
     case ConnectionReferenceKeyFormat.ServiceProvider:
@@ -465,6 +513,20 @@ const serializeHost = (
         )
       );
   }
+};
+
+const mergeHostWithInputs = (hostInfo: Record<string, any>, inputs: any): any => {
+  for (const [key, value] of Object.entries(hostInfo)) {
+    if (inputs[key]) {
+      // eslint-disable-next-line no-param-reassign
+      inputs[key] = { ...inputs[key], ...value };
+    } else {
+      // eslint-disable-next-line no-param-reassign
+      inputs[key] = value;
+    }
+  }
+
+  return inputs;
 };
 
 //#endregion