+++ libs/services/designer-client-services/src/lib/consumption/manifests/xml.ts
@@ -115,7 +115,6 @@ export const xmlTransformManifest = {
           'x-ms-editor': 'dropdown',
           'x-ms-editor-options': {
             multiselect: true,
-            valueSeparator: ',',
             titleSeparator: ',',
             items: [
               {